move `internal/router/router.go` to `internal/router.go`
